грязная имплементация парсинга контрактов

+++ src/moex_publications_bot.cr
@@ -30,24 +30,55 @@ class Main
     parser = Moex::Parser.new(timeout: @@browser_timeout.to_i, filepath: @@file_path, ff_address: @@ff_address)
     poller = Moex::Poller.new(send_last: @@send_last)
     bot = Telegram::PublisherBot.new(@@chat_id, @@api_key)
+    contracts_shown = false
+    subscribe = poller.subscribe_news
 
-    subscribe = poller.subscribe
     loop do
-      Log.info { "Start polling moex..." }
-      obj = parser.go(subscribe.receive)
-      begin
-        bot.send_moex_update obj
-        Log.info { "Sent message => #{obj}" }
-      rescue e
-        Log.info { "Can't send: #{e.message}" }
+      time = Time.local(location: Time::Location.load("Europe/Moscow"))
+      Log.info { "#{time.to_s} Start polling moex..." }
+
+      # #### показываем эти контракты по будням после 9:30 MSK #####
+      if self.time_to_show_contracts? && !contracts_shown
+        Log.info { "Time to show contracts..." }
+        %w(BR-4.20 RTS-3.20 SI-3.20).each do |contract|
+          begin
+            obj = parser.parse_contract(contract)
+            bot.send_moex_update obj unless obj.nil?
+            images = obj[:images].as(Array(String))
+            unless images.empty?
+              Log.info { "Deleting stale screenshots #{images.join(", ")}" }
+              images.each { |image| File.delete image }
+            end
+          rescue
+            next
+          end
+        end
+        contracts_shown = true
+      elsif !self.time_to_show_contracts?
+        contracts_shown = false
       end
 
-      images = obj[:images].as(Array(String))
-      unless images.empty?
-        Log.info { "Deleting stale screenshots #{images.join(", ")}" }
-        images.each { |image| File.delete image }
+      # #### тут дальше парсим только новости #####
+      obj = parser.parse_news(subscribe.receive)
+      unless obj.nil?
+        begin
+          bot.send_moex_update obj
+          Log.info { "Sent message => #{obj}" }
+        rescue e
+          Log.info { "Can't send: #{e.message}" }
+        end
+        images = obj[:images].as(Array(String))
+        unless images.empty?
+          Log.info { "Deleting stale screenshots #{images.join(", ")}" }
+          images.each { |image| File.delete image }
+        end
       end
     end
+  end
+
+  def self.time_to_show_contracts?
+    current_time = Time.local(location: Time::Location.load("Europe/Moscow"))
+    current_time.hour >= 9 && current_time.minute >= 30 && current_time.day_of_week.to_s.in? %w(Monday Tuesday Wednesday Thursday Friday)
   end
 end
 
@@ -55,5 +86,5 @@ begin
   Main.configure
   Main.run
 rescue e
-  puts "Runtime failed: #{e.message}"
+  puts "Runtime restarted with #{e.message}"
 end
